cast m_channel to int with 'cout'

+++ tsMuxer/aacStreamReader.cpp
@@ -8,7 +8,7 @@ const std::string AACStreamReader::getStreamInfo()
 {
     std::ostringstream str;
     str << "Sample Rate: " << m_sample_rate / 1000 << "KHz  ";
-    str << "Channels: " << m_channels;
+    str << "Channels: " << static_cast<int>(m_channels);
     return str.str();
 }
